Fixed warnings.

+++ src/Mobs/AggressiveMonster.cpp
@@ -108,14 +108,14 @@ void cAggressiveMonster::Attack(float a_Dt)
 
 bool cAggressiveMonster::IsMovingToTargetPosition()
 {
-	static const float epsilon = 0.000000000001;
+	static const float epsilon = 0.000000000001f;
 	// Difference between destination x and target x is negligible (to 10^-12 precision)
 	if (fabsf((float)m_FinalDestination.x - (float)m_Target->GetPosX()) < epsilon)
 	{
 		return false;
 	}
 	// Difference between destination z and target z is negligible (to 10^-12 precision)
-	else if (fabsf(m_FinalDestination.z - (float)m_Target->GetPosZ()) > epsilon)
+	else if (fabsf((float)m_FinalDestination.z - (float)m_Target->GetPosZ()) > epsilon)
 	{
 		return false;
 	}
